requirements.txt
boto3==1.4.8
dj-database-url==0.4.2
django==2.0.2
django-debug-toolbar==1.9.1
django-rest-framework==0.1.0
django_storages==1.6.5
djangorestframework==3.7.7
gunicorn==19.7.1
psycopg2==2.7.3.2
whitenoise==3.3.1
